Added: Sending e-mail when connection is restored

+++ uptimem.sh
@@ -3,24 +3,37 @@
 # This is licensed under MIT license.
 # Website: https://github.com/Florke64/uptimem
 
-ID="PARIS"
+ID="$(hostname)"
 HOSTS="$@"
 SUBSCRIBERS=$(cat "subscribers.txt")
 CACHE_DIR="$HOME/.cache/uptimem"
 
-notification() {
-    CURTIME="$(date --iso-8601=seconds)"
-    MAILDIR="/tmp/mail/$CURTIME/"
-    mkdir -p $MAILDIR
-    MAIL="$MAILDIR/notification_message.txt"
-    
-    echo "Subject: $ID: Uptime check failed with $1" > $MAIL
+# E-Mail Content
+SUBJECT_FAIL="Subject: $ID: Uptime check failed with $1"
+SUBJECT_SUCC="Subject: $ID: Connection with $1 is back up"
+BODY_FAIL="You will be informed when it is back up via separate e-mail."
+BODY_SUCC="Remote machine answered the ping."
 
-    echo "You will be informed when it is back up via separate e-mail." >> $MAIL
-    echo "---" >> $MAIL
-    echo "$CURTIME" >> $MAIL
+notification() {
+    status=$2
+    curtime="$(date --iso-8601=seconds)"
+    maildir="/tmp/mail/$curtime/"
+    mkdir -p "$maildir"
+    mail="$maildir/notification_message.txt"
+
+    if [ $status -eq 1 ]
+    then
+        echo "$SUBJECT_FAIL" > $mail
+        echo "$BODY_FAIL" >> $mail
+    else    
+        echo "$SUBJECT_SUCC" > $mail
+        echo "$BODY_SUCC" >> $mail
+    fi
     
-    sendmail $SUBSCRIBERS < $MAIL
+    echo "---" >> $mail
+    echo "$curtime" >> $mail
+    
+    sendmail "$SUBSCRIBERS" < $mail
 }
 
 test() {
@@ -33,28 +46,28 @@ test() {
         echo "Basic test failed."
         test "$1" 4 && return 1
     fi
-    
+
     return 1
 }
 
 main() {
     for host in $HOSTS; do
         lockfile="$CACHE_DIR/$host"
-    
+
         if test "$host" 1
         then
             echo "Ping succeess!"
-            rm -f "$lockfile"
+            if ls "$lockfile" > /dev/null; then rm -f "$lockfile" && notification "$host" 0; fi
         else
             ls $CACHE_DIR > /dev/null || mkdir -p $CACHE_DIR
-            
+
             if ls "$lockfile" > /dev/null
             then
                 echo "Uptime check with $host failed once again."
             else
                 echo "[!] Uptime check with $host failed."
                 touch "$lockfile"
-                notification $host
+                notification "$host" 1
             fi
         fi
     done
